B2B Test Improvements
