+++ cypress-shared/integration/b2b/2.13.2-sales_manager_scenarios.spec.js
@@ -4,23 +4,19 @@ import {
 } from '../../support/common/support.js'
 import b2b from '../../support/b2b/constants.js'
 import {
-  ROLE_DROP_DOWN,
   ROLE_ID_EMAIL_MAPPING as roleObject,
-  STATUSES,
+  ROLE_DROP_DOWN,
 } from '../../support/b2b/utils.js'
 import { loginToStoreFront } from '../../support/b2b/login.js'
 import {
-  productShouldNotbeAvailableTestCase,
   salesUserShouldImpersonateNonSalesUser,
   userShouldNotImpersonateThisUser,
   verifySession,
   stopImpersonation,
 } from '../../support/b2b/common.js'
 import {
-  searchQuote,
   createQuote,
-  updateQuote,
-  filterQuoteByStatus,
+  searchQuote,
   quoteShouldbeVisibleTestCase,
   quoteShouldNotBeVisibleTestCase,
 } from '../../support/b2b/quotes.js'
@@ -30,7 +26,12 @@ function QuotesAccess(
   organizationB,
   organizationBQuote
 ) {
-  quoteShouldNotBeVisibleTestCase(
+  quoteShouldbeVisibleTestCase(
+    organizationName,
+    quotes.OrganizationAdmin.quotes1,
+    organizationName
+  )
+  quoteShouldbeVisibleTestCase(
     organizationName,
     quotes.Buyer2.quotes1,
     organizationName
@@ -40,75 +41,55 @@ function QuotesAccess(
     organizationBQuote.OrganizationAdmin.quotes1,
     organizationB
   )
-  quoteShouldbeVisibleTestCase(
-    organizationName,
-    quotes.OrganizationAdmin.quotes1,
-    organizationName
-  )
 }
 
-describe('Organization A - Cost Center A1 - Sales Rep Scenario', () => {
+describe('Organization A - Cost Center A1 - Sales Manager Impersonation Scenario', () => {
   loginViaCookies({ storeFrontCookie: false })
 
-  const {
-    nonAvailableProduct,
-    users,
-    product,
-    costCenter1,
-    quotes,
-    gmailCreds,
-  } = b2b.OrganizationA
+  const { users, product, costCenter1, gmailCreds } = b2b.OrganizationA
 
   const { organizationName: organizationB, quotes: organizationBQuote } =
     b2b.OrganizationB
 
-  const impersonatedRole = ROLE_DROP_DOWN.Buyer
-
   loginToStoreFront(
-    users.SalesRep,
-    roleObject.SalesRepresentative.role,
+    users.SalesManager,
+    roleObject.SalesManager.role,
     gmailCreds
   )
   verifySession(
     b2b.OrganizationA,
     costCenter1.name,
-    roleObject.SalesRepresentative.role
+    roleObject.SalesManager.role
   )
-
-  productShouldNotbeAvailableTestCase(nonAvailableProduct)
-  QuotesAccess(b2b.OrganizationA, organizationB, organizationBQuote)
   userShouldNotImpersonateThisUser(
-    roleObject.SalesRepresentative.role,
     roleObject.SalesManager.role,
-    users.SalesManager
+    roleObject.SalesRepresentative.role,
+    users.SalesRep
   )
+  QuotesAccess(b2b.OrganizationA, organizationB, organizationBQuote)
+
+  const impersonatedRole = ROLE_DROP_DOWN.OrganizationAdmin
 
   salesUserShouldImpersonateNonSalesUser(
-    roleObject.SalesRepresentative.role,
+    roleObject.SalesManager.role,
     impersonatedRole,
-    users.Buyer1
+    users.OrganizationAdmin1
   )
+  const quote = 'IMPERSONATE_QUOTE_2'
 
-  searchQuote(quotes.SalesRep.updateQuote)
-  const price = '30.00'
-
-  updateQuote(quotes.SalesRep.updateQuote, { price })
-  filterQuoteByStatus(STATUSES.revised)
   QuotesAccess(b2b.OrganizationA, organizationB, organizationBQuote)
-
-  const quote = 'IMPERSONATE_QUOTE_3'
 
   createQuote({
     product,
     quoteEnv: quote,
-    role: roleObject.SalesRepresentative.role,
+    role: roleObject.SalesManager.role,
     impersonatedRole,
   })
-  searchQuote(quote, users.Buyer1)
+  searchQuote(quote, users.OrganizationAdmin1)
   stopImpersonation(
     b2b.OrganizationA,
     costCenter1.name,
-    roleObject.SalesRepresentative.role
+    roleObject.SalesManager.role
   )
 
   preserveCookie()
