+++ cypress-shared/integration/b2b/2.2.2-add_buyer_approver_organizationA.spec.js
@@ -3,23 +3,11 @@ import {
   preserveCookie,
 } from '../../support/common/support.js'
 import b2b from '../../support/b2b/constants.js'
-import {
-  addPaymentTermsCollectionPriceTablesTestCase,
-  setOrganizationIdInJSON,
-} from '../../support/b2b/common.js'
-import {
-  addAddressinCostCenter,
-  deleteCostCenter,
-  addCostCenter,
-  updateCostCenter,
-  updatePaymentTermsinCostCenter,
-} from '../../support/b2b/cost_center.js'
 import { loginToStoreFront } from '../../support/b2b/login.js'
 import { addUser } from '../../support/b2b/add_users.js'
 import {
   ROLE_DROP_DOWN,
   ROLE_DROP_DOWN_EMAIL_MAPPING as role,
-  PAYMENT_TERMS,
 } from '../../support/b2b/utils.js'
 import {
   createQuote,
@@ -35,7 +23,6 @@ describe('OrganizationA - Create a Buyer and Approver, associate Cost Center and
     organizationName,
     costCenter1,
     costCenter2,
-    costCenter3,
     users,
     product,
     quotes,
@@ -49,27 +36,6 @@ describe('OrganizationA - Create a Buyer and Approver, associate Cost Center and
     ROLE_DROP_DOWN.OrganizationAdmin,
     gmailCreds
   )
-
-  setOrganizationIdInJSON(organizationName, costCenter1.name)
-  addPaymentTermsCollectionPriceTablesTestCase(b2b.OrganizationA)
-
-  // CostCenter 2 - Scenarios
-  addCostCenter(organizationName, costCenter2.name, costCenter2.addresses[0])
-  updatePaymentTermsinCostCenter(
-    organizationName,
-    costCenter2.name,
-    PAYMENT_TERMS.Promissory
-  )
-  addAddressinCostCenter(costCenter2.name, costCenter2.addresses[1])
-
-  // // Cost Center 3 - Scenarios
-  addCostCenter(
-    organizationName,
-    costCenter3.temporaryName,
-    costCenter3.addresses[0]
-  )
-  updateCostCenter(costCenter3.temporaryName, costCenter3.name)
-  deleteCostCenter(costCenter3.name)
 
   addUser({
     organizationName,
