--- cypress-shared/integration/b2b/2.4.2-show_organization_by_adding_prinicipal_sales_channel.spec.js
@@ -1,34 +1,0 @@
-/* eslint-disable jest/valid-expect */
-import {
-  loginViaCookies,
-  preserveCookie,
-} from '../../support/common/support.js'
-import { ROLE_DROP_DOWN } from '../../support/b2b/utils.js'
-import b2b from '../../support/b2b/constants.js'
-import {
-  verifySalesChannel,
-  addBindingsWhichShowsOrganization,
-} from '../../support/b2b/graphql.js'
-import { loginToStoreFront } from '../../support/b2b/login.js'
-
-const { users, gmailCreds } = b2b.OrganizationA
-
-describe('Add binding which shows Organization in profile page', () => {
-  loginViaCookies({ storeFrontCookie: false })
-
-  addBindingsWhichShowsOrganization()
-
-  verifySalesChannel(1)
-
-  loginToStoreFront(
-    users.OrganizationAdmin1,
-    ROLE_DROP_DOWN.OrganizationAdmin,
-    gmailCreds
-  )
-
-  it('Verify Organization is showing up', () => {
-    cy.organizationShouldShowInProfile()
-  })
-
-  preserveCookie()
-})
