+++ cypress-shared/integration/b2b/2.4.3-show_organization_by_adding_principal_sales_channel.spec.js
