+++ cypress-shared/integration/b2b/2.5.1-buyer_scenarios_orgA_costA1.spec.js
@@ -7,7 +7,6 @@ import {
   ROLE_DROP_DOWN_EMAIL_MAPPING as role,
   ROLE_DROP_DOWN,
   ROLE_ID_EMAIL_MAPPING as roleObject,
-  STATUSES,
 } from '../../support/b2b/utils.js'
 import { loginToStoreFront } from '../../support/b2b/login.js'
 import {
@@ -17,16 +16,11 @@ import {
   userShouldNotImpersonateThisUser,
 } from '../../support/b2b/common.js'
 import {
-  createQuote,
   searchQuote,
-  filterQuoteByStatus,
+  createQuote,
   quoteShouldbeVisibleTestCase,
   quoteShouldNotBeVisibleTestCase,
 } from '../../support/b2b/quotes.js'
-import {
-  quickOrderByOneByOneTestCase,
-  quickOrderByOneByOneNegativeTestCase,
-} from '../../support/quick-order/testcase.js'
 
 describe('Organization A - Cost Center A1 - Buyer Scenarios', () => {
   loginViaCookies({ storeFrontCookie: false })
@@ -37,7 +31,6 @@ describe('Organization A - Cost Center A1 - Buyer Scenarios', () => {
     costCenter1,
     users,
     product,
-    product2,
     quotes,
     gmailCreds,
   } = b2b.OrganizationA
@@ -88,29 +81,5 @@ describe('Organization A - Cost Center A1 - Buyer Scenarios', () => {
     quoteEnv: quotes.Buyer.quotes3,
     role: ROLE_DROP_DOWN.Buyer,
   })
-  createQuote({
-    product,
-    quoteEnv: quotes.Buyer.quotes4,
-    role: ROLE_DROP_DOWN.Buyer,
-  })
-  createQuote({
-    product,
-    quoteEnv: quotes.Buyer.quotes5,
-    role: ROLE_DROP_DOWN.Buyer,
-  })
-  quickOrderByOneByOneTestCase(
-    ROLE_DROP_DOWN.Buyer,
-    product2,
-    quotes.Buyer.quotes6,
-    '$486.00'
-  )
-  quickOrderByOneByOneNegativeTestCase(
-    ROLE_DROP_DOWN.Buyer,
-    product2,
-    quotes.Buyer.quotes7,
-    '$24,300.00'
-  )
-  filterQuoteByStatus(STATUSES.ready)
-
   preserveCookie()
 })
