+++ cypress-shared/integration/b2b/2.6.2-buyer_scenarios_orgA_costA2.spec.js
@@ -5,15 +5,8 @@ import {
 import b2b from '../../support/b2b/constants.js'
 import { ROLE_DROP_DOWN, STATUSES } from '../../support/b2b/utils.js'
 import { loginToStoreFront } from '../../support/b2b/login.js'
-import {
-  productShouldNotbeAvailableTestCase,
-  verifySession,
-} from '../../support/b2b/common.js'
-import {
-  createQuote,
-  filterQuoteByStatus,
-  quoteShouldNotBeVisibleTestCase,
-} from '../../support/b2b/quotes.js'
+import { verifySession } from '../../support/b2b/common.js'
+import { createQuote, filterQuoteByStatus } from '../../support/b2b/quotes.js'
 import {
   quickOrderByCategory,
   quickOrderByCategoryNegativeTestCase,
@@ -21,32 +14,10 @@ import {
 
 describe('Organization A - Cost Center A2 - Buyer Scenarios', () => {
   loginViaCookies({ storeFrontCookie: false })
-  const {
-    organizationName,
-    nonAvailableProduct,
-    users,
-    costCenter2,
-    quotes,
-    product3,
-    gmailCreds,
-  } = b2b.OrganizationA
-
-  const { organizationName: organizationB, quotes: organizationBQuote } =
-    b2b.OrganizationB
+  const { users, costCenter2, quotes, product3, gmailCreds } = b2b.OrganizationA
 
   loginToStoreFront(users.Buyer2, ROLE_DROP_DOWN.Buyer, gmailCreds)
   verifySession(b2b.OrganizationA, costCenter2.name, ROLE_DROP_DOWN.Buyer)
-  productShouldNotbeAvailableTestCase(nonAvailableProduct)
-  quoteShouldNotBeVisibleTestCase(
-    organizationName,
-    quotes.OrganizationAdmin.quotes1,
-    organizationName
-  )
-  quoteShouldNotBeVisibleTestCase(
-    organizationName,
-    organizationBQuote.OrganizationAdmin.quotes1,
-    organizationB
-  )
   quickOrderByCategory(ROLE_DROP_DOWN.Buyer, quotes.Buyer2.quotes1, '$84.60')
   quickOrderByCategoryNegativeTestCase(
     ROLE_DROP_DOWN.Buyer,
