+++ cypress-shared/support/b2b/checkout.js
@@ -52,9 +52,14 @@ export function fillContactInfo() {
 export function verifyAddress(address) {
   it('Verify Auto fill Address in checkout', updateRetry(3), () => {
     cy.setorderFormDebugItem()
+    if (cy.state('runnable')._currentRetry > 1) cy.reload()
     cy.get('body').then(($shipping) => {
       if ($shipping.find(selectors.OpenShipping).length) {
         cy.get(selectors.OpenShipping, { timeout: 5000 }).click()
+      }
+
+      if ($shipping.find(selectors.EditShipping).length) {
+        cy.get(selectors.EditShipping).should('be.visible').click()
       }
     })
